Rename `testutil` to `functest`.
