+++ src/rinq/internal/functest/peer.go
@@ -1,4 +1,4 @@
-package testutil
+package functest
 
 import (
 	"sync"
