+++ src/rinq/internal/functest/pkg.go
@@ -1,0 +1,2 @@
+// Package functest contains utilities for writing functional tests for Rinq.
+package functest
